+++ src/util/util.h
@@ -49,9 +49,9 @@ _Noreturn void error(const char *fmt, ...);
 
 void show_error_line(const char *line, const char *p, int len);
 
-bool is_im8(intptr_t x);
-bool is_im16(intptr_t x);
-bool is_im32(intptr_t x);
+bool is_im8(int64_t x);
+bool is_im16(int64_t x);
+bool is_im32(int64_t x);
 const char *skip_whitespaces(const char *s);
 const char *block_comment_start(const char *p);
 const char *block_comment_end(const char *p);
